% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export}
\alias{export}
\title{export}
\usage{
export(func, file, res2 = 300, ratio = 1.5, type = "pdf")
}
\arguments{
\item{func}{function that generates a plot}

\item{file}{output figure file name}

\item{res2}{resolution}

\item{ratio}{ratio between with and height}

\item{type}{pdf, png or tif}
}
\description{
export figures in the required file format (png,tif,pdf)
}
\examples{
EXAMPLE 1
x <- 1:10
y <- 1:10
export(func=plot(x,y),'test1',type='pdf')
export(func=plot(x,y),'test1',type='tif')
export(func=plot(x,y),'test1',type='png')

EXAMPLE 2
plot_function <- function {
par(mfrow=c(1,2))
plot(x,y) 
boxplot(x)
}
export(func=plot_function(),'test2',type='pdf')
export(func=plot_function(),'test2',type='tif')
export(func=plot_function(),'test2',type='png') 
}
\keyword{export,}
\keyword{pdf}
\keyword{png,}
\keyword{tif,}
