% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kml2.R
\name{kml2}
\alias{kml2}
\title{kml2}
\usage{
kml2(obj, altitude, colour, heading, size, size_scale = c(0.5, 2), icon,
  filename = "text.kml", legend_icon = NULL)
}
\arguments{
\item{obj}{spatialpointsdataframe with the parameters}

\item{altitude}{altitude column in the data frame of obj}

\item{colour}{column that represents the colours in the data frame of obj}

\item{heading}{column that represents the heading in the data frame of obj}

\item{size}{variable that represents the size in the data frame of obj}

\item{size_scale}{rescaling min and max value}

\item{icon}{image weblink readible by plotKLM or an image you want to use as icon}

\item{filename}{output file name of the kml}

\item{legend_icon}{png that represents the legend}
}
\description{
Generates a kml including the heading
}
\examples{
prepare a dataframe 
obj <- data.frame(altitude=1:100,colour=heat.colors(100),heading=1:100,variable=runif(100),longitude=seq(15.202020, 16.202020,length.out=100), latitude=seq(45.459069,46.543523,length.out=100))
coordinates(obj) <- c('longitude','latitude')
proj4string(obj)<- CRS("+init=epsg:4326")
Example using png plucked from the internet 
kml2(obj,altitude='altitude',colour='colour',heading='heading',size='variable',icon="http://maps.google.com/mapfiles/kml/shapes/airports.png",filename='test.kml')
Example using png locally stored
# first generate an icon
icon <- function(pch=21){
par(bg=NA)
plot(1,1,axes=FALSE,ann=FALSE, cex =30, col='black',bg='white',pch=pch)
} 
export(func=icon(),'icon',type='png')
# use function with the generated icon
kml2(obj,altitude='altitude',colour='colour',heading='heading',size='variable',icon="icon.png",filename='test.kml')
}
\keyword{heading}
\keyword{kml,}
